+++ src/features/diaries/diarySlice.js
@@ -1,7 +1,7 @@
 import { createSlice } from "@reduxjs/toolkit";
 
 const initialState = {
-  diaries: {},
+  diaries: [],
   error: null,
 };
 
@@ -23,8 +23,28 @@ const diarySlice = createSlice({
     },
     addDiary: {
       reducer(state, action) {
-        state.diaries = [...state, action.payload];
-        console.log(state.diaries);
+        //state.diaries = [...state, action.payload];
+        state.diaries.push(action.payload);
+      },
+    },
+    deleteDiary: {
+      reducer(state, action) {
+        const index = state.diaries.findIndex(
+          (diary) => diary.id === action.payload
+        );
+        if (index !== -1) {
+          state.diaries.splice(index, 1);
+        }
+      },
+    },
+    editDiary: {
+      reducer(state, action) {
+        const index = state.diaries.findIndex(
+          (diary) => diary.id === action.payload.id
+        );
+        if (index !== -1) {
+          state.diaries[index] = action.payload;
+        }
       },
     },
   },
@@ -32,5 +52,6 @@ const diarySlice = createSlice({
 
 export const selectCurrentDiaries = (state) => state.diary.diaries;
 export const selectMyDiaries = (state) => state.diary.myDiaries;
-export const { getAllDiaries, getMyDiaries, addDiary } = diarySlice.actions;
+export const { getAllDiaries, getMyDiaries, addDiary, deleteDiary, editDiary } =
+  diarySlice.actions;
 export default diarySlice.reducer;
